[AUTO BATCH PLUGIN] Add compiled_model.cpp, seprate the compiled model related operations from auto_batch.cpp
